Move uniq_chk.py under validation/
